Successful request to iTunes API.
ajaxRequest to iTunes API returns a JSON object with property, feedUrl. This is the url needed where all audio files of specific podcast searched are stored. We make another GET request to feedUrl, which returns an XML file. Our request parses the returned XML to JSON and we've created the file path to the array with all the episodes of target podcast.

This commit also includes tests used to verify data being sent and returned in the request.